requirements.txt
requests>=2.31
pydantic>=2.0
rich>=13.0
python-dotenv>=1.0
beautifulsoup4>=4.12
streamlit>=1.32
